Moving update components into the correct folder
